backend/requirements.txt
flask
flask-cors
python-dotenv
pydantic
google-generativeai
